requirements.txt
Requires...

censusdata
pandas
matplotlib.pyplot
numpy
scipy.stats